refactor: inferred mode in own file
